+++ pnlibrary-bukkit/src/main/kotlin/ru/privatenull/pnlibrary/bukkit/BukkitPlatformAdapter.kt
@@ -460,7 +460,7 @@ class BukkitPlatformAdapter @JvmOverloads constructor(
 
     @EventHandler
     fun onPluginDisable(event: PluginDisableEvent) {
-        library?.events?.close(event.plugin)
+        library?.plugins?.unregisterOwner(event.plugin)
         library?.tasks?.close(event.plugin)
     }